% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{primary_btn}
\alias{primary_btn}
\title{Primary Button}
\usage{
primary_btn(...)
}
\arguments{
\item{...}{inherits parameters from actionButton}
}
\value{
primary action button
}
\description{
Primary Button
}
