% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{secondary_btn}
\alias{secondary_btn}
\title{Secondary Button}
\usage{
secondary_btn(...)
}
\arguments{
\item{...}{inherits parameters from actionButton}
}
\value{
secondary action button
}
\description{
Secondary Button
}
