% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip_element}
\alias{tooltip_element}
\title{Tooltip}
\usage{
tooltip_element(wrapper_id, element, tooltip_text, placement = "top-start")
}
\arguments{
\item{tooltip_text}{tooltip text}

\item{placement}{placement of tooltip}

\item{element_id}{id for tooltip element wrapper}

\item{...}{element inside wrapper}
}
\value{
element with a tooltip wrapper
}
\description{
Tooltip
}
